Move rate_limit to utils
Fixes #128
